requirement.txt
# Python 3.10+
langchain-core>=0.3,<0.4
langchain-openai>=0.3,<0.4
openai>=1.0
tiktoken>=0.7
langchain_google_genai
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0